#Day13 - Callum
- Completed Part1 and Part2 successfully.

Note:
There was an unexpected complication due to folds which do not fully cover the original part of the paper.
Which subtly "worked" by adding an extra empty size to the paper at the bottom
Which given that the problem never specifies how big the paper is could be valid. However, this caused the small example to throw an error. So I corrected.